requirements.txt
pandas
numpy
scikit-learn
openai
dash[diskcache]
plotly
python-dotenv
psutil
multiprocess
dill
dash-bootstrap-components
gunicorn
statsmodels